- testing with gcc.4.4.1-tdm (win32) - axDebug fixes for gcc.4.4.1-tdm - some typos in _pages.h - axInterpolate.h:
rework of the concept of callback between parent <-> child classes, since static methods cannot access non-static class members:

the parent class "myPlugin" instance and a static callback function are passed to the subclass method my_dsp_instance.process(...), then the instance is passed back to a static callback function in the parent class o,O. this allows non-static members to be accessed from the static callback function. a quick solution with c++ templates. not sure how safe it is and how is the performace. declaring both the callback and process as 'inline' should remove the function call penalties..

\rant

test in:
[liteon_os.cpp <--> axInterpolate.h]

- add dspPRNG0 - test noisegen class with fpu
- add _noise.cpp - test noise fx (there was a problem. todo: more debug)